Add ability to estimate physical core usage on Intel KNL
The Intel KNL chips on Stampede 2 have 272 hardware threads on 68 cores which
are reported as 272 cores by the O/S. For most parallel HPC applications the
best way to use these chips is to run one task per core. So if you
view the CPU usage from the point of view of the hardware threads it
looks like the node is only being 25% used, whereas it is probably
optimal. This change generates an estimate of the physical core usage
based on the hardware thread statistics. The estimate is calulated as
follows:

The statistics for the 4 hardware threads on each core are summed. If
the core has 75% or more idle ticks then the top three quarters are ignored. This case
corresponds to three idle hardware threads out of four. If the core
is less than 75% idle then the idle ticks are zeroed and the
rest of the ticks are rescaled to 100% while maintaining their
relative ratios.

Of course this algorithm only generates an estimate of physical
core usage. There are pathalogical cases where it will not correctly
represent the usage. For example, if the code maxes out the four
hardware threads on a core for a quarter of the time then this
code would report 100% core usage whereas it will be closer to 25%
actual. This limitation is somewhat mitigated by doing the rescaling
independently at every time point.