Adding new command line arguments for job filtering
Added two new command line arguments to provide for filtering which jobs get
processed based on a jobs node hours. In this first cut I'm defining small jobs
as any job with less then 10,000 node hours. This is based on historical data on
the size of jobs we've been able to process in parallel vs. those we've had to
process in a single thread.

Also, if both small_jobs and large_jobs are specified then an exception will be
thrown.